+++ Realm/RLMQueryUtil.hpp
@@ -34,7 +34,14 @@ void RLMUpdateQueryWithPredicate(tightdb::Query *query, id predicate, RLMSchema 
 // sort an existing view by the specified property name and direction
 void RLMUpdateViewWithOrder(tightdb::TableView &view, RLMObjectSchema *schema, NSString *property, BOOL ascending);
 
-NSUInteger RLMValidatedColumnIndex(RLMObjectSchema *desc, NSString *columnName);
+// validate and returns a property used for sorting - throw for invalid propty types or names
+RLMProperty *RLMValidatedPropertyForSort(RLMObjectSchema *schema, NSString *propName);
+
+// return column index - throw for invalid column name
+NSUInteger RLMValidatedColumnIndex(RLMObjectSchema *schema, NSString *columnName);
+
+// predicate exception
+NSException *RLMPredicateException(NSString *name, NSString *reason);
 
 // This macro validates predicate format with optional arguments
 #define RLM_VARARG(PREDICATE_FORMAT, ARGS) \