+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 
 ### Enhancements
 
+* Add support for sorting RLMArray properties.
 * Add support for querying for nil object properties.
 * Improve error message when specifying invalid literals when creating or 
   initializing RLMObjects.